factored out minicluster functionality
